% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClassifSummList.R
\name{allClassifSummList}
\alias{allClassifSummList}
\title{Collects Prophage-like and Sloping classification pattern-match information}
\usage{
allClassifSummList(bestMatchList)
}
\arguments{
\item{bestMatchList}{Classifications made with patternMatcher function.}
}
\description{
Collects pattern-match information associated with all contigs classified as Prophage-like or Sloping.
}
\keyword{internal}
