% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenlatGTA_matches.R
\name{allgenlatGTA_matches}
\alias{allgenlatGTA_matches}
\title{Collects gen/lat/GTA classification pattern-match information}
\usage{
allgenlatGTA_matches(best_match_list)
}
\arguments{
\item{best_match_list}{Classifications made with pattern_matcher function. Classifications are stored as the first item in the best_match_list.}
}
\description{
Collects pattern information associated with all contigs predicted as containing gen/lat/GTA transduction events based on the results from the pattern_matcher function.
}
\keyword{internal}
