% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsSummary.R
\name{allPatternMatches}
\alias{allPatternMatches}
\title{Collects pattern-match information for all classifications}
\usage{
allPatternMatches(bestMatchList, classifSummTable)
}
\arguments{
\item{bestMatchList}{Classifications made with patternMatcher function.}

\item{classifSummTable}{Classification summary table with
whole-community:VLP-fraction read coverage ratios calculated.}
}
\value{
List
}
\description{
Collects pattern information associated with all contigs classified as
Prophage-like, Sloping and HighCovNoPattern.
}
\keyword{internal}
