% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allprophagelike_matches.R
\name{allprophagelike_matches}
\alias{allprophagelike_matches}
\title{Collects prophage-like classification pattern-match information}
\usage{
allprophagelike_matches(best_match_list)
}
\arguments{
\item{best_match_list}{Classifications made with pattern_matcher function. Classifications are stored as the first item in the best_match_list.}
}
\description{
Collects pattern information associated with all contigs predicted as containing prophages based on the results from the pattern_matcher function.
}
\keyword{internal}
