% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allProphageLikeClassifs.R
\name{allProphageLikeClassifs}
\alias{allProphageLikeClassifs}
\title{Collects Prophage-like classification pattern-match information}
\usage{
allProphageLikeClassifs(bestMatchList)
}
\arguments{
\item{bestMatchList}{Classifications made with patternMatcher function.}
}
\description{
Collects pattern information associated with all contigs classified as Prophage-like.
}
\keyword{internal}
