% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeResultsFunctions.R
\name{allSlopingClassifs}
\alias{allSlopingClassifs}
\title{Collects Sloping classification pattern-match i nformation}
\usage{
allSlopingClassifs(bestMatchList)
}
\arguments{
\item{bestMatchList}{Classifications made with patternMatcher function.}
}
\value{
List
}
\description{
Collects pattern information associated with all contigs classified as
Sloping in the patternMatcher function.
}
\keyword{internal}
