% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alltransduction_events_summarylist.R
\name{alltransduction_events_summarylist}
\alias{alltransduction_events_summarylist}
\title{Collects prophage-like and gen/lat/GTA transduction classification pattern-match information}
\usage{
alltransduction_events_summarylist(best_match_list)
}
\arguments{
\item{best_match_list}{Classifications made with pattern_matcher function. Classifications are stored as the first item in the best_match_list.}
}
\description{
Collects pattern information associated with all contigs classified as prophage-like or gen/lat/GTA based on the results from the pattern_matcher function.
}
\keyword{internal}
