% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alltransduction_withlowratios_summarylist.R
\name{alltransduction_withlowratios_summarylist}
\alias{alltransduction_withlowratios_summarylist}
\title{Collects Pattern-match information associated with prophage-like and gen/lat/GTA classifications and contigs with low whole-community:VLP-fraction read coverage ratios}
\usage{
alltransduction_withlowratios_summarylist(
  best_match_list,
  summarytable_wratios
)
}
\arguments{
\item{best_match_list}{Classifications made with pattern_matcher function. Classifications are stored as the first item in the best_match_list.}

\item{summarytable_wratios}{Classification summary table with whole-community:VLP-fraction read coverage ratios calculated.}
}
\description{
Collects Pattern information associated with all contigs classified as prophage-like or gen/lat/GTA based on the results from the pattern_matcher function and all contigs with 'none' classifications that have low whole-community:VLP-fraction read coverage ratios.
}
\keyword{internal}
