% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_builder.R
\name{block_builder}
\alias{block_builder}
\title{Block Builder}
\usage{
block_builder(viral_subset, windowsize, minblocksize, maxblocksize)
}
\arguments{
\item{viral_subset}{A subset of the read coverage dataset that pertains only to the contig currently being assessed}

\item{windowsize}{The window size used to re-average read coverage datasets}

\item{minblocksize}{The minimum size of the prophage-like block pattern. Default is 10000 bp.}

\item{maxblocksize}{The maximum size of the prophage-like block pattern. Default is NA.}
}
\description{
Build and translate a block pattern going off the left side, right side and full length of the contig.
}
\keyword{internal}
