% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Block_off_left_translator.R
\name{block_off_left_translator}
\alias{block_off_left_translator}
\title{Block off-left pattern translator}
\usage{
block_off_left_translator(viral_subset, windowsize, blocksize)
}
\arguments{
\item{viral_subset}{A subset of the read coverage dataset that pertains only to the contig currently being assessed}

\item{windowsize}{The window size used to re-average read coverage datasets}

\item{blocksize}{The minimum size of the prophage-like block pattern. Default is 10000 bp.}
}
\description{
Build and translate a block pattern going off the left side of the contig. Stop translating when the pattern left on the contig is 10,000bp. Translate the pattern 2000bp at a time.
}
\keyword{internal}
