% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_off_right_translator.R
\name{block_off_right_translator}
\alias{block_off_right_translator}
\title{Block off-right pattern translator}
\usage{
block_off_right_translator(
  viral_subset,
  windowsize,
  minblocksize,
  maxblocksize
)
}
\arguments{
\item{viral_subset}{A subset of the read coverage dataset that pertains only to the contig currently being assessed}

\item{windowsize}{The window size used to re-average read coverage datasets}

\item{minblocksize}{The minimum size of the prophage-like block pattern. Default is 10000 bp.}

\item{maxblocksize}{The maximum size of the prophage-like block pattern. Default is NA.}
}
\description{
Build and translate a block pattern going off the right side of the contig. Stop translating when the pattern left on the contig is 10,000bp. Translate the pattern 2000bp at a time.
}
\keyword{internal}
