% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockPatternFunctions.R
\name{blockBuilder}
\alias{blockBuilder}
\title{Builds prophage-like block patterns}
\usage{
blockBuilder(
  viralSubset,
  windowSize,
  minBlockSize,
  maxBlockSize,
  searchMethod,
  DirectMaxEval
)
}
\arguments{
\item{viralSubset}{A subset of the read coverage pileup that pertains only to
the contig currently being assessed}

\item{windowSize}{The window size used to re-average read coverage pileups}

\item{minBlockSize}{The minimum size of the prophage-like block pattern.
Default is 10000 bp.}

\item{maxBlockSize}{The maximum size of the prophage-like block pattern.
Default is NA.}

\item{searchMethod}{Search method to use. Either "grid" for the original grid
search or "direct" for DIRECT global optimization.}

\item{DirectMaxEval}{Maximum number of DIRECT evaluations to make.
Default is 100.}
}
\value{
List containing three objects
}
\description{
Build and translate a block pattern going off the left side, right side and
full length of the contig.
}
\keyword{internal}
