% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockheight_optim.R
\name{blockheight_optim}
\alias{blockheight_optim}
\title{Block pattern height optimization}
\usage{
blockheight_optim(best_match_info, oldcoverages, oldcovstep)
}
\arguments{
\item{best_match_info}{The information associated with the current best pattern match. Includes the match score, the min and max y-axis values, the length and the start and stop positions of the pattern}

\item{oldcoverages}{The initial coverage values used for pattern matching}

\item{oldcovstep}{The initial difference between coverage values used for pattern matching}
}
\description{
Optimizes heights of block patterns based on scores of initial pattern matches
}
\keyword{internal}
