% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockpattern_translator.R
\name{blockpattern_translator}
\alias{blockpattern_translator}
\title{Full block pattern-translator}
\usage{
blockpattern_translator(viral_subset, best_match_info, windowsize, pattern)
}
\arguments{
\item{viral_subset}{A subset of the read coverage dataset that pertains only to the contig currently being assessed}

\item{best_match_info}{The information associated with the current best pattern match. Includes the match score, the min and max y-axis values, the length and the start and stop positions of the pattern}

\item{windowsize}{The window size used to re-average read coverage datasets}

\item{pattern}{A vector containing the values associated with the block pattern to be translated across the contig}
}
\description{
Translates full block-pattern across a contig. Translate the pattern 2000bp at a time. Stop translating when the pattern is 5000bp from the end of the contig.
}
\keyword{internal}
