% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockTranslator.R
\name{blockTranslator}
\alias{blockTranslator}
\title{Full block pattern-translator}
\usage{
blockTranslator(viralSubset, bestMatchInfo, windowSize, pattern)
}
\arguments{
\item{viralSubset}{A subset of the read coverage pileup that pertains only to
the contig currently being assessed}

\item{bestMatchInfo}{The information associated with the current best
pattern-match.}

\item{windowSize}{The window size used to re-average read coverage pileups}

\item{pattern}{A vector containing the values associated with the block
pattern}
}
\value{
List
}
\description{
Translates full block-pattern across a contig. Translate the pattern 1000 bp
at a time. Stop translating when the pattern is 5000 bp from the end of the
contig.
}
\keyword{internal}
