% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeWindowSize.R
\name{changeWindowSize}
\alias{changeWindowSize}
\title{Change the read coverage rolling mean window size}
\usage{
changeWindowSize(cleanPileup, windowSize)
}
\arguments{
\item{cleanPileup}{A read coverage dataset that has been cleaned and reformatted by the readcovdf_formatter function}

\item{windowSize}{The number of base pairs to average coverage values over. Options are 100, 500, 1000, or 2000 only!}
}
\description{
Re-averages window sizes of read coverage averages. Start with 100bp windows always. Cannot make window size less than 100bp.
}
\keyword{internal}
