% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cleanNumerics}
\alias{cleanNumerics}
\title{cleanNumerics}
\usage{
cleanNumerics(num, margin = 2)
}
\arguments{
\item{num}{A data frame or tibble.}

\item{margin}{Integer: 1 for rows, 2 for columns (default is 2).}
}
\value{
A cleaned data frame.
}
\description{
Removes rows or columns containing any NA values.
}
