% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanVLPFraction_sampledata.R
\docType{data}
\name{CleanVLPFraction_sampledata}
\alias{CleanVLPFraction_sampledata}
\title{Cleaned VLP-Fraction of Sample Dataset}
\format{
## 'CleanVLPFraction_sampledata'
A data frame with 25,668 rows and 3 columns:
\describe{
 \item{V1}{Contig accession}
 \item{V2}{Read coverage binned over binsize used by pileup.sh}
 \item{V3}{Starting position (bp) of each bin used by pileup.sh. Restarts from 0 at the start of each new contig.}
}
}
\usage{
CleanVLPFraction_sampledata
}
\description{
A subset of contigs from the cleaned VLP-fraction read coverage pileup file generated from BBMap's pileup.sh.
Report...
}
\keyword{datasets}
