% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contig_classification_summary.R
\name{contig_classification_summary}
\alias{contig_classification_summary}
\title{Summarizes pattern-matches into a table}
\usage{
contig_classification_summary(best_match_list)
}
\arguments{
\item{best_match_list}{Classifications made with pattern_matcher function. Classifications are stored as the first item in the best_match_list.}
}
\description{
Summarizes the classications made in the pattern_matcher function based on which pattern match achieved the lowest mean absolute difference for each contig. Outputs results in a table.
}
\keyword{internal}
