% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patternCharacterization.R
\name{contigClassSumm}
\alias{contigClassSumm}
\title{Summarizes pattern-match information}
\usage{
contigClassSumm(bestMatchList)
}
\arguments{
\item{bestMatchList}{Classifications made with patternMatcher function.}
}
\value{
dataframe
}
\description{
Summarizes the classifications made in the patternMatcher() function into a
dataframe.
}
\keyword{internal}
