% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockPatternFunctions.R
\name{directBlockBuilder}
\alias{directBlockBuilder}
\title{DIRECT block pattern optimizer}
\usage{
directBlockBuilder(
  viralSubset,
  windowSize,
  minBlockSize,
  blockLength,
  blockLengthFull,
  minReadCov,
  maxReadCov,
  startingCovs,
  DirectMaxEval,
  globalLocal
)
}
\arguments{
\item{viralSubset}{A subset of the read coverage pileup that pertains only to
the contig currently being assessed}

\item{windowSize}{The window size used to re-average read coverage pileups}

\item{minBlockSize}{The minimum size of the prophage-like block pattern.}

\item{blockLength}{Maximum left and right block pattern length}

\item{blockLengthFull}{Maximum full block pattern length}

\item{minReadCov}{Baseline value outside of the block}

\item{maxReadCov}{Maximum VLP-fraction read coverage value}

\item{startingCovs}{Candidate coverage values used to initialize the search}

\item{DirectMaxEval}{Maximum number of DIRECT evaluations to make. Default is 100.}

\item{globalLocal}{Use global or local DIRECT search. Default is local.}
}
\value{
List containing three objects
}
\description{
Optimizes prophage-like block patterns using DIRECT global optimization.
}
\keyword{internal}
