% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{dmta.asfc}
\alias{dmta.asfc}
\title{dmta.asfc}
\usage{
dmta.asfc(
  sur,
  TmpDir,
  size.x = 256,
  size.y = 256,
  size.n = 256,
  type = "single"
)
}
\arguments{
\item{sur}{surface file}

\item{TmpDir}{temporary folder}

\item{size.x}{total x length of the surface, in pixels}

\item{size.y}{total y length of the surface, in pixels}

\item{size.n}{in case type = "multi", the number of cells for the calculation of heterogeneity}

\item{type}{character vector indicating whether a single value should be calculated for the whole surface ("single"); or if multiple values should be calculated for each n cells ("multi")}
}
\value{
A list containing the .sur file, the type (complexity), Asfc2 and R2adj
}
\description{
Computes Asfc variable
}
\examples{
#to do
tmp <- tempdir()
sur <- system.file("tests/sur/test_surf2.sur", package = "trident")

ResAsf <- dmta.asfc(sur = sur, TmpDir = tmp, size.x = 256, 
                                             size.y = 256, size.n = 256, type = "single")

dplyr::glimpse(ResAsf)

}
