% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmta.R
\name{dmta.spatial}
\alias{dmta.spatial}
\title{dmta.spatial}
\usage{
dmta.spatial(sur, size.x = 256, size.y = 256, size.n = 256, type = "single")
}
\arguments{
\item{sur}{surface file}

\item{size.x}{total x length of the surface, in pixels}

\item{size.y}{total y length of the surface, in pixels}

\item{size.n}{in case type = "multi", the number of cells for the calculation of heterogeneity}

\item{type}{character vector indicating whether a single value should be calculated for the whole surface ("single"); or if multiple values should be calculated for each n cells ("multi")}
}
\value{
the result of fun
}
\description{
Description
}
\examples{
#to do
}
