% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exp_form}
\alias{exp_form}
\title{Format with exponential form}
\usage{
exp_form()
}
\value{
string vector of JavaScript code
}
\description{
A character vector of JavaScript code to format numbers in exponential form.
}
\details{
Format with exponential form
}
