% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{FactorFromFileName}
\alias{FactorFromFileName}
\title{FactorFromFileName}
\usage{
FactorFromFileName(x, separator, position, prefix = "", suffix = "")
}
\arguments{
\item{x}{A character vector of file names.}

\item{separator}{A character string used to split the file names.}

\item{position}{An integer indicating the position of the factor to extract after splitting the file names.}

\item{prefix}{Add a prefix to the factor}

\item{suffix}{Add a suffix to the factor}
}
\value{
A character vector of extracted factors, optionally completed with a prefix and/or a suffix.
}
\description{
Extract a factor from a file name, based on a separator and a position. The extracted factor can be completed with a prefix and/or a suffix.
}
\details{
FactorFromFileName
}
\examples{
# 1. We start with a character vector of file names
files <- c("2023_Albumin_01.csv", "2023_Hemoglobin_02.csv", "2023_Insulin_03.csv")

# 2. We want to split by "_" and extract the second element (between the underscores)
#    We also want to add a prefix "ID_" to the extracted factor
names <- FactorFromFileName(
  x         = files, 
  separator = "_", 
  position  = 2, 
  prefix    = "ID_", 
  suffix    = ""
)

print(names)
# Result : [1] "ID_Albumin" "ID_Hemoglobin" "ID_Insulin"
}
