% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_annotation_search_and_plot.R
\name{gene_annotation_search_and_plot}
\alias{gene_annotation_search_and_plot}
\title{Gene annotation search and plot}
\usage{
gene_annotation_search_and_plot(
  subset_gene_annots,
  keywords,
  viral_subset,
  classification,
  ref_name,
  colindex,
  start_bprange,
  end_bprange
)
}
\arguments{
\item{subset_gene_annots}{The gene annotations associated with the contig being searched}

\item{keywords}{The key-word(s) you would like to search for. Case independent. Searches will return the full gene annotations that contain the matching key-word. Key-word(s) must be in quotes, comma-separated, and surrounded by c() i.e( c("antibiotic", "resistance", "drug") )}

\item{viral_subset}{A subset of the read coverage dataset that pertains only to the contig currently being assessed}

\item{classification}{The classification of the contig being searched}

\item{ref_name}{The reference name for the contig being searched}

\item{colindex}{The column index of either 'gene' or 'product'}

\item{start_bprange}{The basepair at which the search is started if a 'specific' search is used}

\item{end_bprange}{The basepair at which the search is ended if a 'specific' search is used}
}
\description{
Search contigs for gene annotations that match the provided keywords and plot the contigs with matches and their respective locations on the contig
}
\keyword{internal}
