% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneAnnotationSearch.R
\name{geneAnnotationPlot}
\alias{geneAnnotationPlot}
\title{Gene annotation plot}
\usage{
geneAnnotationPlot(
  geneAnnotSubset,
  keywords,
  VLPpileupSubset,
  colIdx,
  startbpRange,
  endbpRange,
  pattern,
  windowSize
)
}
\arguments{
\item{geneAnnotSubset}{Subset of gene annotations to be plotted}

\item{keywords}{The key-word(s) used for the search.}

\item{VLPpileupSubset}{A subset of the pileup associated with the contig/chunk being assessed}

\item{colIdx}{The gff column being searched}

\item{startbpRange}{The basepair at which the search is started if a 'specific' search is used}

\item{endbpRange}{The basepair at which the search is ended if a 'specific' search is used}

\item{pattern}{The pattern-match information associated with the contig/chunk being assessed}

\item{windowSize}{The number of basepairs to average read coverage values over.}
}
\description{
Plot read coverage and location of gene annotations that match the keywords and
search criteria for contig currently being assessed
}
\keyword{internal}
