% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneAnnotationSearch.R
\name{geneSearch}
\alias{geneSearch}
\title{Search for gene annotations on positively classified contigs}
\usage{
geneSearch(
  TrIdentResults,
  VLPpileup,
  gff,
  searchCol,
  keyWords,
  inPatMat = FALSE,
  bpRange = 0,
  saveFilesTo,
  verbose = TRUE
)
}
\arguments{
\item{TrIdentResults}{The output from `TrIdent()`.}

\item{VLPpileup}{A .txt file containing mapped sequencing read coverages averaged over
100 bp windows/bins.}

\item{gff}{A .gff file imported using Bioc.gff::readGFF that is 
associated with the whole-community contigs}

\item{searchCol}{The exact column name (in quotes) in the gff file that you'd like to search for specific 
`keyWords`. Commonly annotation, product, gene, or acc columns, for example, however the specific
name may change based on annotation tool used. For example: "pfam_desc".}

\item{keyWords}{The keyWord(s) to search for. Case independent. Searches will return the string
that contains the matching keyWord. KeyWord(s) must be in quotes, comma-separated, and surrounded by
c() i.e( c("antibiotic", "resistance", "drug") )}

\item{inPatMat}{TRUE or FALSE. If TRUE, only search for gene-annotations in
the pattern-match region. Default is FALSE (i.e search the
entire contig for the gene annotation key-words)}

\item{bpRange}{If `inPatMat` = TRUE, the user may specify the region (in base pairs) that should
be searched to the left and right of the pattern-match region. Default is 0.}

\item{saveFilesTo}{Optional, Provide a path to the directory you wish to save
output to. A folder will be made within the provided directory to store
results.}

\item{verbose}{TRUE or FALSE. Print progress messages to console. Default is TRUE.}
}
\value{
list of ggplot objects
}
\description{
Search contigs classified with TrIdent for gene-annotations that match a provided
key-word(s). Outputs read coverage plots for contigs with matching annotations.
}
\examples{
data("VLPFractionSamplePileup")
data("TrIdentSampleOutput")
data("gffSample")

patternMatches <- geneSearch(
  TrIdentResults = TrIdentSampleOutput,
  VLPpileup = VLPFractionSamplePileup,
  gff = gffSample,
  searchCol="pfam_desc",
  keyWords=c("toxin", "drug", "resistance", "phage")
)

}
