% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff3_cleanup.R
\name{gff3_cleanup}
\alias{gff3_cleanup}
\title{.gff3 file re-formatter}
\usage{
gff3_cleanup(gff_file)
}
\arguments{
\item{gff_file}{A gff3 file containing gene annotation information}
}
\description{
Gene annotations cleanup
}
\keyword{internal}
