% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAremover.R
\name{is.nan.data.frame}
\alias{is.nan.data.frame}
\title{NA remover}
\usage{
\method{is.nan}{data.frame}(x)
}
\arguments{
\item{x}{dataset with potential NAs}
}
\description{
Removes NAs from dataframe. Function found on Stack Exchange.
}
\seealso{
\url{https://stackoverflow.com/questions/18142117/how-to-replace-nan-value-with-zero-in-a-huge-data-frame/18143097#18143097}
}
\keyword{internal}
