% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockPatterns.R
\name{leftRightBlockTranslater}
\alias{leftRightBlockTranslater}
\title{Translate left and right block patterns across contig}
\usage{
leftRightBlockTranslater(
  viralSubset,
  pattern,
  leftOrRight,
  windowSize,
  minReadCov,
  cov,
  bestMatchInfo,
  minBlockSize
)
}
\arguments{
\item{viralSubset}{A subset of the read coverage pileup that pertains only to
the contig currently being assessed}

\item{pattern}{The pattern vector being translated}

\item{leftOrRight}{Is the left or right block pattern being translated}

\item{windowSize}{The window size used to re-average read coverage pileups}

\item{minReadCov}{The baseline value used for the region outside of the block
pattern (either 0 or the minimum VLP-fraction read coverage for the contig)}

\item{cov}{The height value currently being used for the block pattern}

\item{bestMatchInfo}{The information associated with the current best
pattern-match.}

\item{minBlockSize}{The minimum size of the Prophage-like block pattern.
Default is 10,000 bp.}
}
\value{
List
}
\description{
Translates left and right block patterns across contigs 1000 bp at a time
}
\keyword{internal}
