% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBlockPattern.R
\name{makeBlockPattern}
\alias{makeBlockPattern}
\title{Make block patterns for pattern-matching}
\usage{
makeBlockPattern(
  viralSubset,
  windowSize,
  fullLeftRight,
  blockLength,
  nonBlock,
  minReadCov,
  cov
)
}
\arguments{
\item{viralSubset}{A subset of the read coverage pileup that pertains only to the contig currently being assessed}

\item{windowSize}{The window size used to re-average read coverage pileups}

\item{fullLeftRight}{The block pattern variation being built}

\item{blockLength}{Maximum block pattern length}

\item{nonBlock}{Maximum non-block pattern length}

\item{minReadCov}{Either 0 or the minimum VLP-fraction read coverage value}

\item{cov}{The height value of the block pattern}
}
\value{
List containing two objects
}
\description{
Make full, left and right block patterns for Prophage-like classifications
}
\keyword{internal}
