% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullSlopingPatternFunctions.R
\name{makeFullSlopesDirect}
\alias{makeFullSlopesDirect}
\title{Make full slope patterns}
\usage{
makeFullSlopesDirect(leftOrRight, viralSubset, dims, windowSize, minSlope)
}
\arguments{
\item{leftOrRight}{Generate pattern for negative slope (left to right, i.e.
'Left') or positive slope (right to left, i.e. 'Right')}

\item{viralSubset}{A subset of the read coverage pileup that pertains only to
the contig currently being assessed}

\item{dims}{Slope top and bottom values}

\item{windowSize}{The window size used to re-average read coverage pileups}

\item{minSlope}{The minimum slope value to test for sloping patterns}
}
\value{
List
}
\description{
Makes slope patterns sloping either left to right (Left) or right to left
(right) across the contig being assessed using DIRECT dimensions
}
\keyword{internal}
