% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSlopesWStarts.R
\name{makeSlopesWStarts}
\alias{makeSlopesWStarts}
\title{Make slope patterns with starts}
\usage{
makeSlopesWStarts(leftOrRight, viralSubset, newMax, minReadCov, windowSize)
}
\arguments{
\item{leftOrRight}{Generate pattern for negative slope (left to right, i.e. 'Left') or positive slope (right to left, i.e. 'Right')}

\item{viralSubset}{A subset of the read coverage pileup that pertains only to the contig currently being assessed}

\item{newMax}{A value for the top of the sloping pattern that is slightly higher than the maximum coverage value on the viralSubset}

\item{minReadCov}{Minimum read coverage value of the viralSubset}

\item{windowSize}{The window size used to re-average read coverage pileup}
}
\value{
List
}
\description{
Makes slope patterns sloping either left to right (Left) or right to left (right) across the contig being assessed.
Slope patterns contain an initiation point.
}
\keyword{internal}
