% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchsize_checker.R
\name{matchsize_checker}
\alias{matchsize_checker}
\title{Pattern-match size calculator}
\usage{
matchsize_checker(
  classificationsummary,
  transductionclassification_list,
  windowsize
)
}
\arguments{
\item{classificationsummary}{Classification summary table}

\item{transductionclassification_list}{A list containing pattern match information associated with all contigs classified as prophage-like or gen/lat/GTA. Generated with the alltransduction_events_summarylist}

\item{windowsize}{The window size used to re-average read coverage datasets}
}
\description{
Calculate the size, in base pairs, of the matching region for prophage-like and gen/lat/GTA patterns
}
\keyword{internal}
