% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbialToViralRatioCalc.R
\name{microbialToViralRatioCalc}
\alias{microbialToViralRatioCalc}
\title{Whole-community:VLP-fraction read coverage ratio calculator}
\usage{
microbialToViralRatioCalc(classifSummTable, WCpileup, VLPpileup)
}
\arguments{
\item{classifSummTable}{Classification summary table}

\item{WCpileup}{A table containing contig names, coverages averaged over 100 bp windows, and contig positions associated with mapping whole-community reads to whole-community contigs}

\item{VLPpileup}{A table containing contig names, coverages averaged over 100 bp windows, and contig positions associated with mapping VLP-fraction reads to whole-community contigs}
}
\description{
Calculate the whole-community:VLP-fraction read coverage ratio for every contig using the median
read coverage values. If the ratio is less than 2 (i.e VLP-fraction read coverage
is, on average, at least half the whole-community read coverage), then the contig is classified as HighCovNoPattern
}
\keyword{internal}
