% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noPattern.R
\name{noPattern}
\alias{noPattern}
\title{No pattern pattern-match}
\usage{
noPattern(viralSubset, searchMethod, DirectMaxEval)
}
\arguments{
\item{viralSubset}{A subset of the read coverage pileup that pertains only to
the contig currently being assessed}

\item{searchMethod}{Search method to use. Either "grid" for the original grid
search or "direct" for DIRECT global optimization.}

\item{DirectMaxEval}{Maximum number of DIRECT evaluations to make.}
}
\value{
List
}
\description{
A horizontal line at the mean or median coverage should be an optimal
pattern-match if the contig read coverage displays no sloping or block
patterns
}
\keyword{internal}
