% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notransduction_pattern.R
\name{notransduction_pattern}
\alias{notransduction_pattern}
\title{No transduction pattern match}
\usage{
notransduction_pattern(viral_subset)
}
\arguments{
\item{viral_subset}{A subset of the read coverage dataset that pertains only to the contig currently being assessed}
}
\description{
Assess whether a contig does not have a read coverage pattern associated with a transduction event. A horizontal line at the mean coverage should be an optimal match if the contig read coverage displays no transduction patterns
}
\keyword{internal}
