% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noPattern.R
\name{optimNoPattern}
\alias{optimNoPattern}
\title{Function for applying DIRECT global optimization algorithm}
\usage{
optimNoPattern(viralSubset, DirectMaxEval, globalLocal)
}
\arguments{
\item{viralSubset}{A subset of the read coverage pileup that pertains only to
the contig currently being assessed}

\item{DirectMaxEval}{Maximum number of DIRECT evaluations to make.}

\item{globalLocal}{Use global or local DIRECT search. Default is local.}
}
\value{
List
}
\description{
Set upper and lower limits for DIRECT and apply it to the noPattern pattern
}
\keyword{internal}
