% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialSlopingPatternFunctions.R
\name{optimSlopeWithStart}
\alias{optimSlopeWithStart}
\title{Function for applying DIRECT global optimization algorithm}
\usage{
optimSlopeWithStart(
  minSlope,
  minSlopeSize,
  viralSubset,
  windowSize,
  leftOrRight,
  DirectMaxEval,
  globalLocal
)
}
\arguments{
\item{minSlope}{The minimum slope value to test for sloping patterns}

\item{minSlopeSize}{The minimum width of sloping patterns.}

\item{viralSubset}{A subset of the read coverage pileup that pertains only to
the contig currently being assessed}

\item{windowSize}{The window size used to re-average read coverage pileups}

\item{leftOrRight}{Generate pattern for negative slope (left to right, i.e.
'Left') or positive slope (right to left, i.e. 'Right')}

\item{DirectMaxEval}{Maximum number of DIRECT evaluations to make.}

\item{globalLocal}{Use global or local DIRECT search. Default is local.}
}
\value{
List object
}
\description{
Set upper and lower limits for DIRECT and apply it to slope top, bottom, and start values
}
\keyword{internal}
