% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_builder.R
\name{pattern_builder}
\alias{pattern_builder}
\title{Pattern-builder}
\usage{
pattern_builder(
  viral_subset,
  transductionclassifications_wlowratios,
  classification,
  i
)
}
\arguments{
\item{viral_subset}{A subset of the read coverage dataset that pertains only to the contig currently being assessed}

\item{transductionclassifications_wlowratios}{A list containing pattern match information associated with all contigs classified as prophage-like, gen/lat/GTA, or none with high VLP-fraction:whole-community read coverage ratios. Generated with the alltransduction_withlowratios_summarylist}

\item{classification}{The classification of the contig assigned by the main TrIdent_classifier function}

\item{i}{The index position associated with the current contig's best pattern-match information}
}
\description{
Builds the pattern (vector) associated with the 'best pattern-match' for each contig classified as prophage-like, gen/lat/GTA, or none with high VLP:WC read coverage
}
\keyword{internal}
