% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_pattern_matcher.R
\name{pattern_matcher}
\alias{pattern_matcher}
\title{Main pattern-matching function}
\usage{
pattern_matcher(
  phageread_dataset,
  microbialread_dataset,
  windowsize,
  minblocksize,
  maxblocksize
)
}
\arguments{
\item{phageread_dataset}{A table containing contig names, coverages averaged over 100bp windows, and contig positions associated with mapping VLP-fraction reads to whole-community contigs}

\item{microbialread_dataset}{A table containing contig names, coverages averaged over 100bp windows, and contig positions associated with mapping whole-community reads to whole-community contigs}

\item{windowsize}{The window size used to re-average read coverage datasets}

\item{minblocksize}{The minimum size of the prophage-like block pattern. Default is 10000 bp.}

\item{maxblocksize}{The maximum size of the prophage-like block pattern. Default is NA}
}
\description{
Creates the viral_subset, representative of one contig, that is used as input for each individual pattern-building function. After the information associated with the best match for each pattern is obtained, the pattern with the lowest mean absolute difference (score) is chosen as the prediction for the contig being assessed. Prior to the pattern-building and translating, contigs less than 30,000bp and contigs where the value 50th from the max (5000bp) is less than or equal to 10 are removed. Contigs less than 45,000bp are matched against all three prophage patterns and the 'no transduction' pattern. Contigs greater than 45,000bp but less than 100,000 are matched against all three prophage patterns, the two generalized patterns with slopes that start on the contig, and the 'no transduction' pattern. Contigs greater than 100,000bp are matched against all pattern varieties.
}
\keyword{internal}
