% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patternBuilder.R
\name{patternBuilder}
\alias{patternBuilder}
\title{Pattern-builder}
\usage{
patternBuilder(viralSubset, classifList, classification, i)
}
\arguments{
\item{viralSubset}{A subset of the read coverage pileup that pertains only to the contig currently being assessed}

\item{classifList}{A list containing pattern match information associated with all classified contigs.}

\item{classification}{The contig's classification assigned by the TrIdentClassifier function}

\item{i}{The list index associated with each contig's pattern-match information}
}
\description{
Builds the pattern (vector) associated with the 'best pattern-match' for each
contig classified as Prophage-like, Sloping, or HighCovNoPattern.
}
\keyword{internal}
