% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_bivariate}
\alias{plot_bivariate}
\title{plot_bivariate}
\usage{
plot_bivariate(pca_result, factors, factor_col, pc_x = 1, pc_y = 2)
}
\arguments{
\item{pca_result}{A PCA object (from FactoMineR).}

\item{factors}{A data frame containing categorical variables for the individuals.}

\item{factor_col}{String. The name of the column in `factors` to use for grouping.}

\item{pc_x}{Integer. The principal component to plot on the X-axis (default: 1).}

\item{pc_y}{Integer. The principal component to plot on the Y-axis (default: 2).}
}
\value{
A ggplot2 object.
}
\description{
Bivariate PCA Plot with Active and Supplementary Individuals
}
\details{
plot_bivariate
}
\examples{
data(decathlon2, package = "factoextra")

res.pca <- FactoMineR::PCA(
    decathlon2[1:23, 1:12],
    quanti.sup = 11:12,
    ind.sup    = 20:23,
    graph      = FALSE)


factors <- decathlon2[ "Competition"]

factor_col <- "Competition"

plot_bivariate(
  pca_result = res.pca,
  factors = factors,
  factor_col = "Competition",
  pc_x = 1,
  pc_y = 2
)
}
