% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_screeplot}
\alias{plot_screeplot}
\title{plot_screeplot}
\usage{
plot_screeplot(pca_result)
}
\arguments{
\item{pca_result}{A PCA result object returned by \code{FactoMineR::PCA}.}
}
\value{
A \code{ggplot2} object representing the scree plot.
}
\description{
Generates a scree plot (eigenvalues bar plot) from a PCA result object
produced by \code{FactoMineR::PCA}. The plot shows the percentage of
variance explained by each principal component.
}
\details{
Plot Scree Plot from PCA Result
}
\examples{
data(iris)

res_pca <- FactoMineR::PCA(iris[, 1:4], graph = FALSE)

plot_screeplot(res_pca)
}
