% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_TrIdentPatternMatches.R
\name{Plot_TrIdentPatternMatches}
\alias{Plot_TrIdentPatternMatches}
\title{Plot read coverage graphs of contigs classified as Prophage-like, Gen/Lat/GTA, or HighVLPWCRC}
\usage{
Plot_TrIdentPatternMatches(
  VLP_pileup,
  WC_pileup,
  transductionclassifications,
  matchscorefilter,
  SaveFilesTo,
  cleanup = TRUE
)
}
\arguments{
\item{VLP_pileup}{A dataframe containing contig names, coverages averaged over 100bp windows, and contig positions associated with mapping VLP-fraction reads to whole-community contigs}

\item{WC_pileup}{A dataframe containing contig names, coverages averaged over 100bp windows, and contig positions associated with mapping whole-community reads to whole-community contigs}

\item{transductionclassifications}{Output from TrIdent_Classifier}

\item{matchscorefilter}{Filter the results using a pattern match threshold. The pattern match threshold can be defined using the density plot output with TrIdent_Classifier()}

\item{SaveFilesTo}{Provide a path to the directory you wish to save TrIdent plots to. TrIdent will create a folder within the provided directory to store results. This parameter is useful if running TrIdent on the command line.}

\item{cleanup}{If pileup files are generated by BBMap's pileup.sh, the files output from pileup.sh can be input directly into TrIdent with cleanup=TRUE. If your pileup files are NOT generated with BBMap's pileup.sh, set cleanup=FALSE. The user is responsible for data cleaning/formatting if this option is used.}
}
\description{
Plot the read coverages of a contig and its associated pattern-match for each contig classified as prophage-like, gen/lat/GTA or none with high VLP-fraction:whole-community read coverage ratios. Returns a list of plot objects.
}
\examples{
\dontrun{
pattern_matches <- Plot_TrIdentPatternMatches(VLP_pileup=VLP_fracreadcov, WC_pileup=whole_commreadcov, transductionclassifications=TrIdent_results)
pattern_matches$NODE_1
}
}
