% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{polynom.sur}
\alias{polynom.sur}
\title{polynom.sur}
\usage{
polynom.sur(sur, TmpDir, deg = 8, path = TRUE)
}
\arguments{
\item{sur}{surface file}

\item{TmpDir}{temporary folder}

\item{deg}{the polynomial degree to be subtracted to the surface}

\item{path}{logical, indicates if the path of the cleaned surface should be returned as a character string in R}
}
\value{
a cleaned surface, with a path to its current location if path = TRUE (default)
}
\description{
A function to remove a given polynomial from a surface
}
\details{
polynom.sur
}
\examples{
tmp <- tempdir()

sur <- system.file("tests/sur/test_surf.sur", package = "trident")

out <- file.path(tmp, "test_surf.sur")

file.copy(from = sur, to = out)

sur <- out

ResPol <- polynom.sur(sur = sur, TmpDir = tmp, deg = 1, path = TRUE)
}
