% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{prepare_pca_data}
\alias{prepare_pca_data}
\title{prepare_pca_data}
\usage{
prepare_pca_data(df, var_order = NULL)
}
\arguments{
\item{df}{Data frame to process.}

\item{var_order}{Vector of column names to reorder (or NULL).}
}
\value{
A list with components numerics and factors.
}
\description{
Prepare data for PCA by separating numeric and factor columns and optionally reordering columns.
}
\details{
prepare_pca_data
}
\examples{
df <- data.frame(a = 1:3, b = factor(c("x","y","z")), c = 4:6)
prepare_pca_data(df)
prepare_pca_data(df, var_order = c("c", "a"))
}
