% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophagelike_border_finder.R
\name{prophagelike_border_finder}
\alias{prophagelike_border_finder}
\title{Prophage-like border finder}
\usage{
prophagelike_border_finder(
  viral_subset,
  transductionclassificationpatterns,
  i,
  windowsize
)
}
\arguments{
\item{viral_subset}{A subset of the read coverage dataset that pertains only to the contig currently being assessed}

\item{transductionclassificationpatterns}{The pattern match information associated with each contig classified as prophage-like, gen/lat/GTA, or none with high VLP-fraction:whole-community read coverage ratios}

\item{i}{The index for the contig currently being assessed}

\item{windowsize}{The window size used to re-average read coverage datasets}
}
\description{
Find borders of prophage-like patterns with more specificity than pattern-matching
}
\keyword{internal}
