% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophageLikeActivity.R
\name{prophageLikeActivity}
\alias{prophageLikeActivity}
\title{Prophage-like classification activity/abundance}
\usage{
prophageLikeActivity(
  classifSummTable,
  prophageLikeClassifList,
  VLPpileup,
  WCpileup,
  windowSize
)
}
\arguments{
\item{classifSummTable}{Classification summary table}

\item{prophageLikeClassifList}{A list containing pattern match information associated with all contigs classified as prophage-like. Generated with the allprophagelike_matches}

\item{VLPpileup}{A table containing contig names, coverages averaged over 100bp windows, and contig positions associated with mapping VLP-fraction reads to whole-community contigs}

\item{WCpileup}{A table containing contig names, coverages averaged over 100bp windows, and contig positions associated with mapping whole-community reads to whole-community contigs}

\item{windowSize}{The window size used to re-average read coverage datasets}
}
\value{
Dataframe
}
\description{
Determines whether a detected prophage-like element has read coverage in the whole-community that is either elevated or depressed compared to the average read coverage of the non-prophage region.
}
\keyword{internal}
