% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophagelikeactivity_checker.R
\name{prophagelikeactivity_checker}
\alias{prophagelikeactivity_checker}
\title{Prophage-like activity/abundance checker}
\usage{
prophagelikeactivity_checker(
  classificationsummary,
  prophageclassifications,
  phageread_dataset,
  microbialread_dataset,
  windowsize
)
}
\arguments{
\item{classificationsummary}{Classification summary table}

\item{prophageclassifications}{A list containing pattern match information associated with all contigs classified as prophage-like. Generated with the allprophagelike_matches}

\item{phageread_dataset}{A table containing contig names, coverages averaged over 100bp windows, and contig positions associated with mapping VLP-fraction reads to whole-community contigs}

\item{microbialread_dataset}{A table containing contig names, coverages averaged over 100bp windows, and contig positions associated with mapping whole-community reads to whole-community contigs}

\item{windowsize}{The window size used to re-average read coverage datasets}
}
\description{
Determines whether a detected prophage/prophage-like element is active/highly abundant based on elevated (>1.2x the average read coverage of the non-prophage-like region) read coverage of the prophage-like region in the whole-community reads.
}
\keyword{internal}
