% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophageLikeBorders.R
\name{prophageLikeBorders}
\alias{prophageLikeBorders}
\title{Prophage-like border finder}
\usage{
prophageLikeBorders(viralSubset, classificationPatterns, i, windowSize)
}
\arguments{
\item{viralSubset}{A subset of the read coverage pileup that pertains only to the contig currently being assessed}

\item{classificationPatterns}{The pattern match information associated with each contig classified as Prophage-like, Sloping, or HighCovNoPattern}

\item{i}{The index for the contig currently being assessed}

\item{windowsize}{The window size used to re-average read coverage pileups}
}
\description{
Find borders of Prophage-like patterns with more specificity than pattern-matching using
100 bp window pileups and sliding standard deviation technique.
}
\keyword{internal}
