% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readcovdf_formatter.R
\name{readcovdf_formatter}
\alias{readcovdf_formatter}
\title{Correctly formats input read coverage summary files.}
\usage{
readcovdf_formatter(read_dataset)
}
\arguments{
\item{read_dataset}{A table containing contig names, coverages averaged over 100bp windows, and contig positions}
}
\description{
Places columns in correct order and renames columns. Clean the contig labels to remove excess informatio.
}
\keyword{internal}
