% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{renderDistributionPlot}
\alias{renderDistributionPlot}
\title{Render a distribution plot}
\usage{
renderDistributionPlot(df, x_var, y_var, type = "violin")
}
\arguments{
\item{df}{A data frame containing the variables to plot.}

\item{x_var}{A character string specifying the name of the categorical variable
to be mapped on the x-axis (grouping variable).}

\item{y_var}{A character string specifying the name of the numeric variable
to be mapped on the y-axis (response variable).}

\item{type}{A character string specifying the type of plot to produce.
Must be either `"violin"` (default) or `"boxplot"`.}
}
\value{
A ggplot object representing the distribution plot.
}
\description{
This function creates a distribution plot (either a violin plot or a boxplot) for a specified data frame and variables. The user can choose the type of plot and customize the colors.
}
\details{
Render a distribution plot (violin or boxplot) for a given data frame and specified variables.
}
\examples{
data(iris)

renderDistributionPlot(
  df    = iris,
  x_var = "Species",
  y_var = "Sepal.Length",
  type  = "violin"
)
}
