% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{run_pca}
\alias{run_pca}
\title{run_pca}
\usage{
run_pca(numerics, active_cols, supp_vars = NULL, supp_inds = NULL)
}
\arguments{
\item{numerics}{data.frame of the active numeric variables.}

\item{active_cols}{Names of the active columns selected for the PCA.}

\item{supp_vars}{Names of the supplementary variables (character vector, may be NULL).}

\item{supp_inds}{data.frame of supplementary individuals (may be NULL).}
}
\value{
PCA object (FactoMineR) containing the results.
}
\description{
Run a PCA using the active numeric variables, with optional supplementary
variables (quanti.sup) and supplementary individuals (ind.sup). The function
assembles the data matrix according to the presence of these extras and then
calls FactoMineR::PCA.
}
\details{
run_pca
}
\examples{
data(iris)

# numeric columns
iris_num <- iris[, 1:4]

# active columns
train <- iris_num[1:100, ]

# supplementary individuals
supp_inds <- iris_num[101:110, ]

# PCA
res <- run_pca(
  numerics    = train,
  active_cols = c("Sepal.Length", "Sepal.Width"),
  supp_vars   = c("Petal.Length", "Petal.Width"),
  supp_inds   = supp_inds
)

# results
summary(res)
}
