% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_LefttoRight.R
\name{slope_LefttoRight}
\alias{slope_LefttoRight}
\title{Sloping pattern left-to-right pattern builder}
\usage{
slope_LefttoRight(viral_subset, windowsize)
}
\arguments{
\item{viral_subset}{A subset of the read coverage dataset that pertains only to the contig currently being assessed}

\item{windowsize}{The window size used to re-average read coverage datasets}
}
\description{
Build a sloping pattern that consists of a sloping line spanning the contig being assessed. The line slopes from left to right. The slope of the line is changed, but the pattern is not translated across the contig.
}
\keyword{internal}
