% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_LefttoRight_withstart.R
\name{slope_LefttoRight_withstart}
\alias{slope_LefttoRight_withstart}
\title{Build sloping pattern that slopes left-to-right with an initial jump-up in read coverage}
\usage{
slope_LefttoRight_withstart(viral_subset, windowsize)
}
\arguments{
\item{viral_subset}{A subset of the read coverage dataset that pertains only to the contig currently being assessed}

\item{windowsize}{The window size used to re-average read coverage datasets}
}
\description{
Build a sloping pattern that consists of a sloping line that starts on the contig being assessed. The pattern is used as input for the slopepattern_translator. The slope of the pattern is changed after each full translation across a contig.
}
\keyword{internal}
