% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_summary.R
\name{slope_summary}
\alias{slope_summary}
\title{Summarize slopes for Gen/Lat/GTA classifications}
\usage{
slope_summary(classificationsummary, genlatGTAclassifications)
}
\arguments{
\item{classificationsummary}{Classification summary table}

\item{genlatGTAclassifications}{A list containing pattern match information associated with all contigs classified as gen/lat/GTA. Generated with the allgenlatGTA_matches}
}
\description{
Add slope information for Gen/Lat/GTA classifications to summary table
}
\keyword{internal}
