% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopepattern_translator.R
\name{slopepattern_translator}
\alias{slopepattern_translator}
\title{Sloping pattern translator}
\usage{
slopepattern_translator(
  viral_subset,
  best_match_info,
  windowsize,
  pattern,
  direction
)
}
\arguments{
\item{viral_subset}{A subset of the read coverage dataset that pertains only to the contig currently being assessed}

\item{best_match_info}{The information associated with the current best pattern match. Includes the match score, the min and max y-axis values, the length and the start and stop positions of the pattern}

\item{windowsize}{The window size used to re-average read coverage datasets}

\item{pattern}{A vector containing the values associated with the generalized pattern to be translated across the contig}

\item{direction}{The direction of the slope of the generalized pattern being translated across. Either "lefttoright" or "righttoleft"}
}
\description{
Translates a sloping pattern containing the initial jump-up in read coverage across a contig. Translate the pattern 2000bp at a time. Stop translating when the pattern left on the contig reaches 45,000bp.
}
\keyword{internal}
