% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialSlopingPatternFunctions.R
\name{slopeWithStartDirect}
\alias{slopeWithStartDirect}
\title{Sloping pattern with an initial jump-up in read coverage}
\usage{
slopeWithStartDirect(
  viralSubset,
  windowSize,
  minSlope,
  minSlopeSize,
  leftOrRight,
  dims
)
}
\arguments{
\item{viralSubset}{A subset of the read coverage pileup that pertains only to
the contig currently being assessed}

\item{windowSize}{The window size used to re-average read coverage pileups}

\item{minSlope}{The minimum slope value to test for sloping patterns}

\item{minSlopeSize}{The minimum width of sloping patterns.}

\item{leftOrRight}{Generate pattern for negative slope (left to right, i.e.
'Left') or positive slope (right to left, i.e. 'Right')}

\item{dims}{Slope top, bottom, and start position}
}
\value{
List object
}
\description{
Build, translate, and change slope of sloping pattern with slope start using 
DIRECT dimensions
}
\keyword{internal}
