% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_transduction_search_and_plot.R
\name{spec_transduction_search_and_plot}
\alias{spec_transduction_search_and_plot}
\title{Specialized transduction search and plot}
\usage{
spec_transduction_search_and_plot(
  ref_name,
  phageread_dataset,
  transductionclassifications,
  transductionclassificationsummary,
  windowsize,
  i,
  noreadcov,
  spectranslength
)
}
\arguments{
\item{ref_name}{The reference name of the contig currently being assessed (i.e "NODE_1")}

\item{phageread_dataset}{A table containing contig names, coverages averaged over 100bp windows, and contig positions associated with mapping VLP-fraction reads to whole-community contigs}

\item{transductionclassifications}{The pattern match information associated with each contig classified as prophage-like, gen/lat/GTA, or none with high VLP-fraction:whole-community read coverage ratios}

\item{transductionclassificationsummary}{The summary information associated with each contig classified as prophage-like, gen/lat/GTA, or none with high VLP-fraction:whole-community read coverage ratios}

\item{windowsize}{The window size used to re-average read coverage datasets}

\item{i}{The index for the contig currently being assessed}

\item{noreadcov}{How many bp of no read coverage are encountered before specialized transduction searching stops? Default is 500.}

\item{spectranslength}{How many bp of read coverage are needed for specialized transduction to be considered? Default is 2000.}
}
\description{
Search contigs classified as prophage-like for potential specialized transduction and return the plot visualizing the search results. Plots in green have potential specialized transduction events.
}
\keyword{internal}
