% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specializedTransductionID.R
\name{specializedTransductionID}
\alias{specializedTransductionID}
\title{Identify potential specialized transduction events on contigs classified as Prophage-like}
\usage{
specializedTransductionID(
  VLPpileup,
  TrIdentResults,
  specificContig,
  noReadCov = 500,
  specTransLength = 2000,
  matchScoreFilter = Inf,
  logScale = FALSE,
  cleanup = TRUE
)
}
\arguments{
\item{VLPpileup}{A table containing contig names, coverages averaged over 100 bp windows, and contig positions associated with mapping VLP-fraction reads to whole-community contigs}

\item{TrIdentResults}{Output from TrIdentClassifier}

\item{specificContig}{Provide the name of a specific contig if you would like to search only that contig.i.e. "NODE_1"}

\item{noReadCov}{How many bp of no read coverage are encountered before specialized transduction searching stops? Default is 500.}

\item{specTransLength}{How many bp of read coverage are needed for specialized transduction to be considered? Default is 2000.}

\item{matchScoreFilter}{Filter results to only display contigs that recieved a pattern-match score below the chosen cutoff. Defualt is infinity (i.e no filter)}

\item{logScale}{If TRUE, coverage is plotted in log10. If FALSE, raw coverage values are plotted. Default is FALSE.}

\item{cleanup}{If pileup files are generated by BBMap, the .txt files can be input directly into TrIdent with cleanup=TRUE. The user is responsible for data cleaning/formatting if cleanup=FALSE.}
}
\description{
Search contigs classified as Prophage-like for dense read coverage outside of the pattern-match borders
that may indicate specialized transduction. Returns a list with the first object containing a summary table
and the second object containing a list of plots of with associated specialzied transduction search results.
If the plot is green, it has been identified as having potential specialized transduction.
}
\examples{
data("VLPFractionSamplePileup")
data("TrIdentSampleOutput")

specTransduction <- specializedTransductionID(
                    VLPpileup=VLPFractionSamplePileup,
                    TrIdentResults=TrIdentSampleOutput
                    )

specTransductionNODE62 <- specializedTransductionID(
                          VLPpileup=VLPFractionSamplePileup,
                          TrIdentResults=TrIdentSampleOutput,
                          specificContig="NODE_62"
                          )
}
