% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specializedTransductionHelpers.R
\name{specTransductionPlot}
\alias{specTransductionPlot}
\title{Specialized transduction plot}
\usage{
specTransductionPlot(
  viralSubsetZoom,
  startPosBp,
  endPosBp,
  SpecTransLeft,
  specTransRight,
  contigName,
  classifPatternMatches,
  i,
  specTransSumm,
  logScale,
  classifSumm
)
}
\arguments{
\item{viralSubsetZoom}{contig subset surrounding Prophage-like pattern-match}

\item{startPosBp}{Left border position}

\item{endPosBp}{Right border position}

\item{SpecTransLeft}{End position of spec transduction on left border}

\item{specTransRight}{End position of spec transduction on right border}

\item{contigName}{The reference name of the contig currently being assessed
(i.e "NODE_1")}

\item{classifPatternMatches}{The pattern match information associated with
each contig classified as prophage-like,  sloping, or HighCovNoPattern}

\item{i}{The index for the contig currently being assessed}

\item{specTransSumm}{Results for spec transduction search}

\item{logScale}{If TRUE, coverage is plotted in log10. If FALSE, raw coverage
values are plotted. Default is FALSE.}

\item{classifSumm}{The summary information associated with each contig
classified as Prophage-like, Sloping, or HighCovNoPattern}
}
\value{
ggplot object
}
\description{
Plot search results of `specializedTransductionID()`
}
\keyword{internal}
