% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggThreshold.R
\name{suggThreshold}
\alias{suggThreshold}
\title{Calculates a suggested filtering threshold for plotting/saving TrIdent results}
\usage{
suggThreshold(normMatchScores, histogram)
}
\arguments{
\item{normMatchScores}{Normalized pattern-match score vector}

\item{histogram}{A histogram object created from normalized pattern-match scores}
}
\description{
Uses the distribution of normalized pattern match-scores to calculate a suggested filtering threshold. The
threshold is two standard deviations to the right of the maximum histogram value.
}
\keyword{internal}
