% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sur.map.R
\name{sur.map}
\alias{sur.map}
\title{sur.map}
\usage{
sur.map(sur, col.levels = 255, col.range = NULL, col.type = "temperature")
}
\arguments{
\item{sur}{A surface file of the .SUR file format}

\item{col.levels}{The number of colors that will be used to describe height variation}

\item{col.range}{A vector of colors to be used on the map; default is NULL.}

\item{col.type}{A pre-defined pattern of colors, either "grayscale", "temperature" (default) or "colorblind"}
}
\value{
A data frame or a matrix of the 3d coordinates of the surface
}
\description{
Displays the height map of a .SUR file in an rgl window
}
