% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sur.read.R
\name{sur.read}
\alias{sur.read}
\title{sur.read}
\usage{
sur.read(sur, type = "xyz")
}
\arguments{
\item{sur}{surface file}

\item{type}{character vector indicating the desired output format - either a matrix of heights, or a data.frame with x, y and z as columns}
}
\value{
the result of fun
}
\description{
Description
}
\examples{
#to do
}
