% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrIdent-package.R
\docType{package}
\name{TrIdent-package}
\alias{TrIdent}
\alias{TrIdent-package}
\title{TrIdent: TrIdent- Transduction Identification}
\description{
Automatic detection, classification and characterization of potential active transducing events based on sequencing read coverage patterns.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/jlmaier12/TrIdent/issues}
}

}
\keyword{internal}
