% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{trident.arrange}
\alias{trident.arrange}
\title{trident.arrange}
\usage{
trident.arrange(
  df,
  y,
  by = "hsd.mean.p.value",
  alpha = 0.05,
  byngr = FALSE,
  geomean = FALSE,
  gp.priority = seq_along(levels(y))
)
}
\arguments{
\item{df}{A data frame of numeric values.}

\item{y}{A factor vector.}

\item{by}{Sorting parameter. See documentation for available options.}

\item{alpha}{Significance level (default 0.05).}

\item{byngr}{Logical: sort by the number of significantly different groups?}

\item{geomean}{Logical: use geometric mean for p-values?}

\item{gp.priority}{Numeric vector to reorder factor levels.}
}
\value{
A data frame of variables ranked by the selected parameter.
}
\description{
Computes multiple statistical tests (ANOVA, Kruskal, Post-hoc) to rank variables.
}
\examples{
library(dplyr)

df <- iris[, 1:4]
y  <- iris$Species

res <- trident.arrange( df    = df,
                        y     = y,
                        by    = "f.stat",
                        alpha = 0.05
                      )
head(res)
}
