% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{trident.boxcox}
\alias{trident.boxcox}
\title{trident.boxcox}
\usage{
trident.boxcox(df, y, list.lambda = FALSE)
}
\arguments{
\item{df}{A data.frame of numeric values.}

\item{y}{A factor representing the grouping variable.}

\item{list.lambda}{Logical. If TRUE, the list of estimated lambdas is
included in the results. Defaults to FALSE.}
}
\value{
A list containing:
  \item{boxcox}{A data.frame of transformed values.}
  \item{lambda}{(optional) A numeric vector of estimated lambdas.}
}
\description{
Transforms the variables of a data.frame using the Box-Cox formula,
  estimating the optimal lambda parameter for each variable via ANOVA.
}
\examples{
csv <- system.file("tests/csv/Example_BoxCox.csv", package = "trident")

df <- read.csv(file = csv, header = TRUE, as.is = TRUE, row.names = NULL)

fact <- as.factor(c(rep("A", 15), rep("B", 15), rep("C", 15)))

for (i in seq_len(ncol(df))) {

    stats::aov(df[, i] ~ fact)$residuals -> residuals

    p.value <- stats::shapiro.test(residuals)$p.value
    
    print(p.value)
}

df_trans  <- trident.boxcox(df = df, y = fact, list.lambda = TRUE)$boxcox
df_lambda <- trident.boxcox(df = df, y = fact, list.lambda = TRUE)$lambda

for (i in seq_len(ncol(df))) {

    stats::aov(df_trans[, i] ~ fact)$residuals -> residuals

    p.value <- stats::shapiro.test(residuals)$p.value
    
    print(c(p.value, df_lambda[i]))
}

}
