% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{trident.hetero}
\alias{trident.hetero}
\title{trident.hetero}
\usage{
trident.hetero(x, var.name = "X")
}
\arguments{
\item{x}{A numeric vector}

\item{var.name}{A character string defining the name of the variable from which heterogeneity should be computed; default 'X'}
}
\value{
A data frame of parameters assessing heterogeneity
}
\description{
Heterogeneity of dmta variables
}
\examples{

x <- c(
  12, 15, NA, 18, 21, -14, NA, 19, 23, 17,
  16, 20, NA, 22, 24,  13, 11, NA, 25, 26,
  27, 28, 29, NA, 30, -10,  9,  8,  7,  6
)

DFstat <- trident.hetero(x, var.name = "ExampleVariable")

dplyr::glimpse(DFstat)

}
