% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Help.R
\docType{package}
\name{TrIdent}
\alias{TrIdent-package}
\alias{TrIdent}
\title{TrIdent- Transduction Identification info}
\description{
Automatic detection, classification and characterization of potential active transducing events based on sequencing read coverage patterns.
}
\author{
Jessie Maier \email{jlmaier@ncsu.edu}
}
\keyword{internal}
