% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{trident.top3}
\alias{trident.top3}
\title{trident.top3s}
\usage{
trident.top3(df, y)
}
\arguments{
\item{df}{A data frame of numeric values.}

\item{y}{A factor vector.}
}
\value{
A list containing the top 3 variable names and a ranked data frame.
}
\description{
Ranks the top 3 variables capable of discriminating categories for each pair of factor levels.
}
\examples{
library(foreach)
library(dplyr)

df <- mtcars[, c("mpg","disp","hp","drat","wt","qsec")]
y  <- factor(mtcars$cyl)

res <- trident.top3(df = df, y = y)

res$ranked[c("group", "aov.p.value")]

}
