% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrIdent_Classifier.R
\name{TrIdent_Classifier}
\alias{TrIdent_Classifier}
\title{Classify contigs as Prophage-like, Gen/Lat/GTA, None, or HighVLPWCratio}
\usage{
TrIdent_Classifier(phageread_dataset, microbial_readdataset, windowsize = 1000)
}
\arguments{
\item{phageread_dataset}{A table containing contig names, coverages averaged over 100bp windows, and contig positions associated with mapping VLP-fraction reads to whole-community contigs}

\item{microbial_readdataset}{A table containing contig names, coverages averaged over 100bp windows, and contig positions associated with mapping whole-community reads to whole-community contigs}

\item{windowsize}{The number of base pairs to average coverage values over. Options are 100, 200, 500, 1000 and 2000 ONLY. Larger window sizes improve processing time but the resolution of read coverage patterns may be lost. Default is 1000bp.}
}
\description{
Classifies contigs as prophage-like, gen/lat/GTA, none, or highVLPWCratio. This function performs all the pattern-matching and summarizes the results into a list. The first item in the list is a table consisting of the summary information of all the contigs that passed through pattern-matching (i.e were not filtered out). The second item in the list is a table consisting of the summary information of all contigs that were predicted as containing a potential prophage, generalized transduction and/or low whole-community: VLP-fraction read coverage ratios. The third item in the list contains the best pattern-match information associated with each contig in the previous table. The fourth and final object in the list is a table containing the contigs that were filtered out prior to pattern_matching and the reason why.
}
\examples{
\dontrun{
TrIdent_results <- TrIdent_Classifier(VLP_fracreadcov, whole_commreadcov, windowsize=1000)
}
}
