% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viral_subset_zoom.R
\name{viral_subset_zoom}
\alias{viral_subset_zoom}
\title{Prophage-like zoom-in}
\usage{
viral_subset_zoom(
  viral_subset,
  transductionclassificationpatterns,
  i,
  zoom,
  windowsize
)
}
\arguments{
\item{viral_subset}{A subset of the read coverage dataset that pertains only to the contig currently being assessed}

\item{transductionclassificationpatterns}{The pattern match information associated with each contig classified as prophage-like, gen/lat/GTA, or none with high VLP-fraction:whole-community read coverage ratios}

\item{i}{The index for the contig currently being assessed}

\item{zoom}{The number of rows outside the start and stop positions of the prophage-like pattern to zoom-in on}

\item{windowsize}{The window size used to re-average read coverage datasets}
}
\description{
Subset contigs with prophage-like patterns to 'zoom-in' on desired region surrounding pattern.
}
\keyword{internal}
