% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VLPFraction_sampledata.R
\docType{data}
\name{VLPFraction_sampledata}
\alias{VLPFraction_sampledata}
\title{VLP-Fraction of Sample Dataset}
\format{
## 'VLPFraction_sampledata'
A data frame with 25,668 rows and 4 columns:
\describe{
 \item{V1}{Contig accession}
 \item{V2}{Read coverage binned over binsize used by pileup.sh}
 \item{V3}{Starting position (bp) of each bin used by pileup.sh. Restarts from 0 at the start of each new contig.}
 \item{V4}{Starting position (bp) of each bin used by pileup.sh. Does NOT restart at the start of each new contig.}
}
}
\usage{
VLPFraction_sampledata
}
\description{
A subset of contigs from the raw VLP-fraction read coverage pileup file generated from BBMap's pileup.sh.
Report...
}
\keyword{datasets}
