% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WCVF_ratio_calculator.R
\name{WCVF_ratio_calculator}
\alias{WCVF_ratio_calculator}
\title{Whole-community:VLP-fraction read coverage ratio calculator}
\usage{
WCVF_ratio_calculator(
  classificationsummary,
  microbialread_dataset,
  phageread_dataset
)
}
\arguments{
\item{classificationsummary}{Classification summary table}

\item{microbialread_dataset}{A table containing contig names, coverages averaged over 100bp windows, and contig positions associated with mapping whole-community reads to whole-community contigs}

\item{phageread_dataset}{A table containing contig names, coverages averaged over 100bp windows, and contig positions associated with mapping VLP-fraction reads to whole-community contigs}
}
\description{
Calculate the whole-community:VLP-fraction read coverage ratio using the median coverage for all contigs. If the ratio is less than 2 (i.e VLP-fraction read coverage is on average at least half the whole-community read coverage), then mark the contig as having a low whole-community (WC) to VLP-fraction (VLPF) read coverage ratio.
}
\keyword{internal}
