% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windowsize_func.R
\name{windowsize_func}
\alias{windowsize_func}
\title{Change the read coverage window size}
\usage{
windowsize_func(read_dataset, windowsize)
}
\arguments{
\item{read_dataset}{A read coverage dataset that has been cleaned and reformatted by the readcovdf_formatter function}

\item{windowsize}{The number of base pairs to average coverage values over. Options are 100, 500, 1000, or 2000 only!}
}
\description{
Re-averages window sizes of read coverage averages. Start with 100bp windows always. Cannot make window size less than 100bp.
}
\keyword{internal}
