% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroCountSearch.R
\name{zeroCountSearch}
\alias{zeroCountSearch}
\title{Counts zero values to the left and right of prophage-like borders}
\usage{
zeroCountSearch(startOrEnd, viralSubsetZoom, startOrEndPosRow, noReadCov)
}
\arguments{
\item{startOrEnd}{searching the start (left side) or end (right side) of the prophage-like pattern match}

\item{viralSubsetZoom}{viralSubset dataframe subsetted to 50,000 bp outside the pattern match borders}

\item{startOrEndPosRow}{The row index of the start or end position of the prophage-like pattern match}

\item{noReadCov}{How many bp of no read coverage are encountered before specialized transduction searching stops? Default is 500.}
}
\description{
Checks to see at which point the number of consecutive zero values to the left and right of the prophage-like pattern match borders equals the noReadCov parameter
}
\keyword{internal}
