% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exData}
\alias{exData}
\title{exData}
\format{
An object of class \code{data.frame} with 60 rows and 17 columns.
}
\usage{
exData
}
\description{
An example of microwear dataset
}
\keyword{datasets}
