% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullSlope.R
\name{fullSlope}
\alias{fullSlope}
\title{Sloping pattern builder}
\usage{
fullSlope(viralSubset)
}
\arguments{
\item{viralSubset}{A subset of the read coverage pileup that pertains only to the contig currently being assessed}
}
\description{
Build a sloping pattern that consists of a sloping line spanning the contig being assessed. The line slopes from left to right. The slope of the line is changed, but the pattern is not translated across the contig.
}
\keyword{internal}
