% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneAnnotationSearch.R
\name{GeneAnnotationSearch}
\alias{GeneAnnotationSearch}
\title{Search/explore gene annotations of contigs classified as having transduction events}
\usage{
GeneAnnotationSearch(
  transductionclassification,
  phageread_dataset,
  gene_annots,
  geneorproduct,
  windowsize,
  keywords,
  genelocation = "nonspecific",
  bprange = 0,
  specificcontig
)
}
\arguments{
\item{transductionclassification}{Output from TrIdent_Classifier.}

\item{phageread_dataset}{A table containing contig names, coverages averaged over 100bp windows, and contig positions associated with mapping VLP-fraction reads to whole-community contigs}

\item{gene_annots}{GFF3 file cleaned with the gff3_cleanup function}

\item{geneorproduct}{Are you searching for a gene name or gene product? "gene" or "product"}

\item{windowsize}{The window size used to re-average read coverage datasets in TrIdent_Classifier. Default is 1000.}

\item{keywords}{The key-word(s) you would like to search for. Case independent. Searches will return the full gene annotations that contain the matching key-word. Key-word(s) must be in quotes, comma-separated, and surrounded by c() i.e( c("antibiotic", "resistance", "drug") )}

\item{genelocation}{Would you like to search for gene-annotations in a specific region surrounding the location of the predicted transduction event? If yes, "specific", if no "nonspecific". Default is "nonspecific" (i.e the entire contig, including any regions not associated with the matching region of the predicted transduction event, will be searched for the gene annotation key-words)}

\item{bprange}{If you are searching for gene annotations in specific ("specific") locations, you may specify the region (in basepairs) that should be searched to the left and right of the predicted transduction event location on the contig. Default is 0 (if you choose specific and don't set the bprange, it will only search gene annotations within the pattern-match region)}

\item{specificcontig}{Provide the name of a specific contig if you would like to search only that contig.i.e. "NODE_1"}
}
\description{
Search contigs classified as prophage-like, gen/lat/GTA or none with high VLP-fraction:whole-community read coverage ratios for gene-annotations that match a provided key-word(s). Outputs read coverage plots for contigs with matching annotations.
}
\examples{
\dontrun{
Virulence_matches <- GeneAnnotationSearch(TrIdent_results, whole_commreadcov, VLP_fracreadcov, cleaned_gff, "product", windowsize=1000, keywords=c("resistance", "antibiotic", "toxin"), genelocation="specific", bprange=10000)
Virulence_matches$NODE_450

CmR_matches <- GeneAnnotationSearch(TrIdent_results, whole_commreadcov, VLP_fracreadcov, cleaned_gff, "gene", windowsize=1000, keywords=c("cmR"), genelocation="specific", bprange=0,)
CmR_matches$NODE_10
Contig_1_GeneAnnots <- GeneAnnotationSearch(TrIdent_results, whole_commreadcov, VLP_fracreadcov, cleaned_gff, "gene", windowsize=1000, keywords=c("cmR"), genelocation="nonspecific",specificcontig="NODE_1")
}
}
\keyword{internal}
