% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multicheck.R
\name{multicheck}
\alias{multicheck}
\title{multicheck2}
\usage{
multicheck(df, y, alpha = 0.05)
}
\arguments{
\item{df}{A dataframe or a subset of a dataset entirely composed of numeric variables}

\item{y}{A factor}

\item{alpha}{Significance level}
}
\value{
A list of ratios and p-values for all the options which where selected using check.- parameters
}
\description{
Perform various statistical checks
}
\examples{
#to do
}
