% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullSlopingPatternFunctions.R
\name{optimSlopeFull}
\alias{optimSlopeFull}
\title{Function for applying DIRECT global optimization algorithm}
\usage{
optimSlopeFull(minSlope, viralSubset, windowSize, leftOrRight, DirectMaxEval)
}
\arguments{
\item{minSlope}{The minimum slope value to test for sloping patterns}

\item{viralSubset}{A subset of the read coverage pileup that pertains only to
the contig currently being assessed}

\item{windowSize}{The window size used to re-average read coverage pileup}

\item{leftOrRight}{Generate pattern for negative slope (left to right, i.e.
'Left') or positive slope (right to left, i.e. 'Right')}

\item{DirectMaxEval}{Maximum number of DIRECT evaluations to make.}
}
\value{
List
}
\description{
Set upper and lower limits for DIRECT and apply it to slope top and bottom
}
\keyword{internal}
