% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_TrIdentPatternMatches.R
\name{Plot_TrIdentPatternMatches}
\alias{Plot_TrIdentPatternMatches}
\title{Plot read coverage graphs of contigs classified as Prophage-like, Gen/Lat/GTA, or HighVLPWCRC}
\usage{
Plot_TrIdentPatternMatches(
  phageread_dataset,
  microbialread_dataset,
  transductionclassifications,
  cleanup = TRUE
)
}
\arguments{
\item{phageread_dataset}{A table containing contig names, coverages averaged over 100bp windows, and contig positions associated with mapping VLP-fraction reads to whole-community contigs}

\item{microbialread_dataset}{A table containing contig names, coverages averaged over 100bp windows, and contig positions associated with mapping whole-community reads to whole-community contigs}

\item{transductionclassifications}{Output from TrIdent_Classifier}

\item{cleanup}{If pileup files are generated by BBMap's pileup.sh, the files output from pileup.sh can be input directly into TrIdent with cleanup=TRUE. If your pileup files are NOT generated with BBMap's pileup.sh, set cleanup=FALSE. The user is responsible for data cleaning/formatting if this option is used.}
}
\description{
Plot the read coverages of a contig and its associated pattern-match for each contig classified as prophage-like, gen/lat/GTA or none with high VLP-fraction:whole-community read coverage ratios. Returns a list of plot objects.
}
\examples{
\dontrun{
pattern_matches <- Plot_TrIdentPatternMatches(VLP_fracreadcov, whole_commreadcov, TrIdent_results)
pattern_matches$NODE_1
}
}
