% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_TrIdentPatternMatches.R
\name{Plot_TrIdentPatternMatches}
\alias{Plot_TrIdentPatternMatches}
\title{Plot read coverage graphs of contigs classified as Prophage-like, Gen/Lat/GTA, or HighVLPWCRC}
\usage{
Plot_TrIdentPatternMatches(
  phageread_dataset,
  microbialread_dataset,
  transductionclassifications,
  windowsize = 1000
)
}
\arguments{
\item{phageread_dataset}{A table containing contig names, coverages averaged over 100bp windows, and contig positions associated with mapping VLP-fraction reads to whole-community contigs}

\item{microbialread_dataset}{A table containing contig names, coverages averaged over 100bp windows, and contig positions associated with mapping whole-community reads to whole-community contigs}

\item{transductionclassifications}{Output from TrIdent_Classifier}

\item{windowsize}{The window size used to re-average read coverage datasets in the TrIdent_Classifier. Default is 1000.}
}
\description{
Plot the read coverages of a contig and its associated pattern-match for each contig classified as prophage-like, gen/lat/GTA or none with high VLP-fraction:whole-community read coverage ratios. Returns a list of plot objects.
}
\examples{
\dontrun{
pattern_matches <- Plot_TrIdentPatternMatches(VLP_fracreadcov, whole_commreadcov, TrIdent_results, windowsize=1000)
pattern_matches$NODE_1
}
}
