% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrIdentResults.R
\name{plotTrIdentResults}
\alias{plotTrIdentResults}
\title{Plot read coverage graphs of contigs classified as Prophage-like, Sloping, or HighCovNoPattern}
\usage{
plotTrIdentResults(
  VLPpileup,
  WCpileup,
  TrIdentResults,
  matchScoreFilter,
  saveFilesTo,
  cleanup = TRUE
)
}
\arguments{
\item{VLPpileup}{A dataframe containing contig names, coverages averaged over 100 bp windows, and contig positions associated with mapping VLP-fraction reads to whole-community contigs}

\item{WCpileup}{A dataframe containing contig names, coverages averaged over 100 bp windows, and contig positions associated with mapping whole-community reads to whole-community contigs}

\item{TrIdentResults}{Output from TrIdentClassifier}

\item{matchScoreFilter}{Filter the results using a pattern match threshold. A suggested filtering threshold is found in the histogram after TrIdentClassifier is run. Default is no filtering.}

\item{saveFilesTo}{Provide a path to the directory you wish to save TrIdent plots to. TrIdent will create a folder within the provided directory to store results.}

\item{cleanup}{If pileup files are generated by BBMap, the .txt files can be input directly into TrIdent with cleanup=TRUE. The user is responsible for data cleaning/formatting if cleanup=FALSE.}
}
\description{
Plot the read coverages of a contig and its associated pattern-match for Prophage-like,
Sloping and HighCovNoPattern classifications. Returns a list of ggplot objects.
}
\examples{
data("VLPFractionSamplePileup")
data("WholeCommunitySamplePileup")
data("TrIdentSampleOutput")
patternMatches <- plotTrIdentResults(
                  VLPpileup=VLPFractionSamplePileup,
                  WCpileup=WholeCommunitySamplePileup,
                  TrIdentResults=TrIdentSampleOutput
                  )
}
