% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmta.R
\name{polynom.sur}
\alias{polynom.sur}
\title{polynom.sur}
\usage{
polynom.sur(sur, deg = 8, path = TRUE, copy = FALSE)
}
\arguments{
\item{sur}{surface file}

\item{deg}{the polynom degree to be substracted to the surface}

\item{path}{logical, indicates if the path of the cleaned surface should be returned as a character string in R}

\item{copy}{logical, indicates if a copy of the cleaned .sur file should be made in the working directory. If FALSE, it remains in a temporary file}
}
\value{
a cleaned surface, with a path to its current location if path = TRUE (default)
}
\description{
A function to remove a given polynom from a surface
}
