% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsHisto.R
\name{resultsHisto}
\alias{resultsHisto}
\title{Create histogram of normalized pattern-match scores}
\usage{
resultsHisto(summaryList)
}
\arguments{
\item{summaryList}{Classification summary table filtered to only include
contigs with Prophage-like, Sloping and HighCovNoPattern classifications}
}
\value{
ggplot object
}
\description{
Plots a histogram of normalized match scores for all Prophage-like, Sloping
and HighCovNoPattern classifications and colors the plot based on the
classifications.
}
\keyword{internal}
