% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopeSumm.R
\name{slopeSumm}
\alias{slopeSumm}
\title{Summarize slopes for sloping classifications}
\usage{
slopeSumm(classifSumm, slopingClassifList, windowSize)
}
\arguments{
\item{classifSumm}{Classification summary table}

\item{slopingClassifList}{A list containing pattern match information
associated with all contigs classified as sloping.}

\item{windowSize}{The window size used to re-average read coverage pileups}
}
\value{
dataframe
}
\description{
Add slope information for sloping classifications to summary table
}
\keyword{internal}
