% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecializedTransduction_ID.R
\name{SpecializedTransduction_ID}
\alias{SpecializedTransduction_ID}
\title{Identify potential specialized transduction events on contigs classified as Prophage-like}
\usage{
SpecializedTransduction_ID(
  phageread_dataset,
  transductionclassification,
  specificcontig,
  noreadcov = 500,
  spectranslength = 2000,
  cleanup = TRUE
)
}
\arguments{
\item{phageread_dataset}{A table containing contig names, coverages averaged over 100bp windows, and contig positions associated with mapping VLP-fraction reads to whole-community contigs}

\item{transductionclassification}{Output from TrIdent_Classifier}

\item{specificcontig}{Provide the name of a specific contig if you would like to search only that contig.i.e. "NODE_1"}

\item{noreadcov}{How many bp of no read coverage are encountered before specialized transduction searching stops? Default is 500.}

\item{spectranslength}{How many bp of read coverage are needed for specialized transduction to be considered? Default is 2000.}

\item{cleanup}{If pileup files are generated by BBMap's pileup.sh, the files output from pileup.sh can be input directly into TrIdent with cleanup=TRUE. If your pileup files are NOT generated with BBMap's pileup.sh, set cleanup=FALSE. The user is responsible for data cleaning/formatting if this option is used.}
}
\description{
Search contigs classified as prophage-like for potential specialized transduction. Returns a list with the first object containing a summary table and the second object containing a list of plots of prophage-like patterns with associated specialzied transduction search results. If the plot is green, it has been identified as having potential specialized transduction.
}
\examples{
\dontrun{
Spec_Transduction <- SpecializedTransduction_ID(VLP_fracreadcov, TrIdent_results)
Summary_Table <- spec_transduction[[1]]
NODE_8_plot <- spec_transduction[[2]]$NODE_8

Spec_Transduction_NODE1 <- SpecializedTransduction_ID(VLP_fracreadcov, TrIdent_results, specificcontig="NODE_1", noreadcov=1000, spectranslength=4000)
}
}
