% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sug_threshold.R
\name{sug_threshold}
\alias{sug_threshold}
\title{Calculates a suggested filtering threshold for plotting/saving TrIdent results}
\usage{
sug_threshold(norm_match_scores, histogram)
}
\arguments{
\item{norm_match_scores}{Mormalized pattern-match score vector}

\item{histogram}{A histogram object created from normaized pattern-match scores}
}
\description{
Uses the distribution of normalized pattern match-scores to calculate a suggested filtering threshold.
}
\keyword{internal}
