% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sur.read.R
\name{sur.read}
\alias{sur.read}
\title{sur.read}
\usage{
sur.read(sur, type = "xyz")
}
\arguments{
\item{sur}{A surface file of the .SUR file format}

\item{type}{A character vector indicating the desired output format - either a matrix of heights, or a data.frame with x, y and z as columns}
}
\value{
A data frame or a matrix of the 3d coordinates of the surface
}
\description{
Read .SUR files
}
