% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trident.arrange.R
\name{trident.arrange}
\alias{trident.arrange}
\title{trident.arrange}
\usage{
trident.arrange(df, y, by = "mean.hsd.p.value", gp.priority = NULL)
}
\arguments{
\item{df}{A vector or a dataframe of numeric values}
}
\value{
the result of fun
}
\description{
Classify variables by their ability to discriminate categories
}
\examples{
#to do
}
