% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{trident.arrange}
\alias{trident.arrange}
\title{trident.arrange}
\usage{
trident.arrange(df, y, by = "p.value", method = "lsd")
}
\arguments{
\item{df}{A vector or a dataframe of numeric values}
}
\value{
the result of fun
}
\description{
Classify variables by their ability to discriminate categories
}
\examples{
#to do
}
