% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{trident.boxcox}
\alias{trident.boxcox}
\title{trident.boxcox}
\usage{
trident.boxcox(df, y, list.lambda = FALSE)
}
\arguments{
\item{df}{A vector or a dataframe of numeric values}

\item{y}{A factor}

\item{list.lambda}{Logical, if TRUE the list of lambda values is added to the results}
}
\value{
A list containing a dataframe of the transformed values
}
\description{
Transforms x according to the Box-Cox formula, using the parameter lambda
}
\details{
to do
}
\examples{
#to do
}
