% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmta.R
\name{trident.hetero}
\alias{trident.hetero}
\title{trident.hetero}
\usage{
trident.hetero(x, var.name = "X")
}
\arguments{
\item{x}{A numeric vector}

\item{var.name}{A character string defining the name of the variable from which heterogeneity should be computed; default 'X'}
}
\value{
A data frame of parameters assessing heterogeneity
}
\description{
Heterogeneity of dmta variables
}
