% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trident.top3.R
\name{trident.top3}
\alias{trident.top3}
\title{trident.arrange}
\usage{
trident.top3()
}
\arguments{
\item{df}{A vector or a dataframe of numeric values}
}
\value{
the result of fun
}
\description{
Classify variables by their ability to discriminate categories
}
\examples{
#to do
}
