% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trident.top3.R
\name{trident.top3}
\alias{trident.top3}
\title{trident.top3}
\usage{
trident.top3(df, y)
}
\arguments{
\item{df}{A vector or a dataframe of numeric values}

\item{y}{A factor}
}
\value{
the result of fun
}
\description{
Classify top 3 variables by their ability to discriminate categories for each group of the factor variable.
}
\examples{
#to do
}
