% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrIdent_Classifier.R
\name{TrIdent_Classifier}
\alias{TrIdent_Classifier}
\title{Classify contigs as Prophage-like, Gen/Lat/GTA, None, or HighVLPWCratio}
\usage{
TrIdent_Classifier(
  VLP_pileup,
  WC_pileup,
  windowsize = 1000,
  minblocksize = 10000,
  maxblocksize = Inf,
  mincontiglength = 30000,
  cleanup = TRUE
)
}
\arguments{
\item{VLP_pileup}{A dataframe containing contig names, coverages averaged over 100bp windows, and contig positions associated with mapping VLP-fraction reads to whole-community contigs}

\item{WC_pileup}{A dataframe containing contig names, coverages averaged over 100bp windows, and contig positions associated with mapping whole-community reads to whole-community contigs}

\item{windowsize}{The number of base pairs to average coverage values over. Options are 100, 200, 500, 1000 and 2000 ONLY. Larger window sizes improve processing time but the resolution of read coverage patterns may be lost. Default is 1000bp.}

\item{minblocksize}{The minimum size of the prophage-like block pattern. Default is 10000 bp.}

\item{maxblocksize}{The maximum size of the prophage-like block pattern. Default is NA}

\item{mincontiglength}{The minimum contig size (in bp) to perform pattern-matching on. Contigs smaller than this threshold will be filtered out. Default is 30,000bp}

\item{cleanup}{If pileup files are generated by BBMap's pileup.sh, the files output from pileup.sh can be input directly into TrIdent with cleanup=TRUE. If your pileup files are NOT generated with BBMap's pileup.sh, set cleanup=FALSE. The user is responsible for data cleaning/formatting if this option is used.}
}
\description{
Classifies contigs as prophage-like, gen/lat/GTA, none, or highVLPWCratio. This function performs all the pattern-matching and summarizes the results into a list. The first item in the list is a table consisting of the summary information of all the contigs that passed through pattern-matching (i.e were not filtered out). The second item in the list is a table consisting of the summary information of all contigs that were predicted as containing a potential prophage, generalized, lateral or GTA (Gen/Lat/GTA) transduction and/or low whole-community: VLP-fraction read coverage ratios. The third item in the list contains the best pattern-match information associated with each contig in the previous table. The fourth object in the list is a table containing the contigs that were filtered out prior to pattern_matching and the reason why. The fifth item is the windowsize used for the search.
}
\examples{
\dontrun{
TrIdent_results <- TrIdent_Classifier(VLP_pileup=VLP_fracreadcov, WC_pileup=whole_commreadcov, windowsize=1000, minblocksize=10000, maxblocksize=150000, cleanup=TRUE)
}
}
