% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrIdentClassifier.R
\name{TrIdentClassifier}
\alias{TrIdentClassifier}
\title{Classify contigs as Prophage-like, Sloping, HighCovNoPattern, and NoPattern}
\usage{
TrIdentClassifier(
  VLPpileup,
  WCpileup,
  windowSize = 1000,
  minBlockSize = 10000,
  maxBlockSize = Inf,
  minContigLength = 30000,
  minSlope = 0.001,
  suggFiltThresh = FALSE,
  SaveFilesTo,
  cleanup = TRUE
)
}
\arguments{
\item{VLPpileup}{A dataframe containing contig names, coverages averaged over 100 bp windows, and contig positions associated with mapping VLP-fraction reads to whole-community contigs.}

\item{WCpileup}{A dataframe containing contig names, coverages averaged over 100 bp windows, and contig positions associated with mapping whole-community reads to whole-community contigs.}

\item{windowSize}{The number of bps to average coverage values over. Options are 100, 200, 500, 1000 ONLY. Larger windowSizes improve processing time but the resolution of read coverage patterns may be lost. Default is 1000 bp.}

\item{minBlockSize}{The minimum size of the Prophage-like block pattern. Default is 10,000 bp.}

\item{maxBlockSize}{The maximum size of the Prophage-like block pattern. Default is NA.}

\item{minContigLength}{The minimum contig size (in bp) to perform pattern-matching on. Must be at least 25,000 bp. Default is 30,000 bp.}

\item{minSlope}{The minimum slope value to test for sloping patterns. Default is 0.001 (i.e minimum change of 10x read coverage over 100,000 bp)}

\item{suggFiltThresh}{TRUE or FALSE, Suggest a filtering threshold on the output pattern-match score histogram. Default is FALSE.}

\item{SaveFilesTo}{Provide a path to the directory you wish to save TrIdent output summary tables to. TrIdent will make a folder within the provided directory to store results.}

\item{cleanup}{If pileup files are generated by BBMap, the .txt files can be input directly into TrIdent with cleanup=TRUE. The user is responsible for data cleaning/formatting if cleanup=FALSE.}
}
\description{
Performs all the pattern-matching and summarizes the results into a list. The first item in the list
is a table consisting of the summary information of all the contigs that passed through pattern-matching
(i.e were not filtered out). The second item in the list is a table consisting of the summary information
of all contigs that were classified via pattern-matching. The third item in the list contains the pattern-match
information associated with each contig in the previous table. The fourth object in the list is a table containing the
contigs that were filtered out prior to pattern-matching. The fifth item is the windowSize used for the search.
}
\examples{
data("VLPFractionSamplePileup")
data("WholeCommunitySamplePileup")
TrIdent_results <- TrIdentClassifier(
                   VLPpileup=VLPFractionSamplePileup,
                   WCpileup=WholeCommunitySamplePileup
                   )
}
